+++ publish.sh
@@ -50,19 +50,49 @@ if [ -z "$ANALYTICS_API_RELEASE_TOKEN" ]; then
     setEnvVariable "ANALYTICS_API_RELEASE_TOKEN" $ANALYTICS_API_RELEASE_TOKEN
 fi
 
-echo "Make sure to bump the libraryVersion and versionCode in the <root>/build.gradle file, README and CHANGELOG first and merge that PR into develop.\n\nAfter releasing, change the \"developLocal\" to false manually and start both the examples and make sure that the outgoing payload doesn't include \"-local\" in the version string (and pull the right version from artifactory).\n"
+echo "Make sure to bump the version in the README and CHANGELOG first and merge that PR into develop.\n\nAfter releasing, change the \"developLocal\" to false manually and start both the examples and make sure that the outgoing payload doesn't include \"-local\" in the version string (and pull the right version from artifactory).\n"
 echo "Version (without leading \"v\")":
 read VERSION
 git checkout develop
 git pull
+
+if ! ./gradlew spotlessCheck --daemon; then
+    echo "Code style violations detected, please fix them first on develop as otherwise the build will fail."
+    exit
+fi
+
+#TODO next release: check if version parameter works (for CircleCI, we also need to add libs-release-local from the .m2/settings.xml, so it will build the other bitmovin and exo collectors (otherwise it can't resolve the dependency on collector before distributing to bintray)
+echo "Creating and publishing :collector project..."
+./gradlew -DdevelopLocal=false -Pversion="$VERSION" :collector:clean || exit
+./gradlew -DdevelopLocal=false -Pversion="$VERSION" :collector:build || exit
+./gradlew -DdevelopLocal=false -Pversion="$VERSION" :collector:assembleRelease || exit
+./gradlew -DdevelopLocal=false -Pversion="$VERSION" :collector:artifactoryPublish || exit
+echo "Created and published :collector project."
+
+echo "Creating and publishing :collector-bitmovin-player project..."
+./gradlew -DdevelopLocal=false -Pversion="$VERSION" :collector-bitmovin-player:clean || exit
+./gradlew -DdevelopLocal=false -Pversion="$VERSION" :collector-bitmovin-player:build || exit
+./gradlew -DdevelopLocal=false -Pversion="$VERSION" :collector-bitmovin-player:assembleRelease || exit
+./gradlew -DdevelopLocal=false -Pversion="$VERSION" :collector-bitmovin-player:artifactoryPublish || exit
+echo "Created and published :collector-bitmovin-player project."
+
+echo "Creating and publishing :collector-exoplayer project..."
+./gradlew -DdevelopLocal=false -Pversion="$VERSION" :collector-exoplayer:clean || exit
+./gradlew -DdevelopLocal=false -Pversion="$VERSION" :collector-exoplayer:build || exit
+./gradlew -DdevelopLocal=false -Pversion="$VERSION" :collector-exoplayer:assembleRelease || exit
+./gradlew -DdevelopLocal=false -Pversion="$VERSION" :collector-exoplayer:artifactoryPublish || exit
+echo "Created and published :collector-exoplayer project."
+
+
 git checkout main
 git pull
 git merge develop
-git tag -a v$VERSION -m "v$VERSION"
+git tag -a v$VERSION -m "v$VERSION" #TODO check if tag exists in the beginning. If yes, ask if the user wants to override the version and force overriding of tag
 git push origin main v$VERSION
-
+#TODO exit if error
 echo "Pushed \"main\" and \"$VERSION\" to repo."
 
+#TODO override existing release
 curl \
   -u bitAnalyticsCircleCi:$ANALYTICS_GH_TOKEN \
   -X POST \
@@ -71,27 +101,6 @@ curl \
   -d "{\"tag_name\":\"v$VERSION\", \"name\": \"v$VERSION\", \"draft\": false}"
 
 echo "Created release in public repo."
-
-echo "Creating and publishing :collector project..."
-./gradlew -DdevelopLocal=false :collector:clean
-./gradlew -DdevelopLocal=false :collector:build
-./gradlew -DdevelopLocal=false :collector:assembleRelease
-./gradlew -DdevelopLocal=false :collector:artifactoryPublish
-echo "Created and published :collector project."
-
-echo "Creating and publishing :collector-bitmovin-player project..."
-./gradlew -DdevelopLocal=false :collector-bitmovin-player:clean
-./gradlew -DdevelopLocal=false :collector-bitmovin-player:build
-./gradlew -DdevelopLocal=false :collector-bitmovin-player:assembleRelease
-./gradlew -DdevelopLocal=false :collector-bitmovin-player:artifactoryPublish
-echo "Created and published :collector-bitmovin-player project."
-
-echo "Creating and publishing :collector-exoplayer project..."
-./gradlew -DdevelopLocal=false :collector-exoplayer:clean
-./gradlew -DdevelopLocal=false :collector-exoplayer:build
-./gradlew -DdevelopLocal=false :collector-exoplayer:assembleRelease
-./gradlew -DdevelopLocal=false :collector-exoplayer:artifactoryPublish
-echo "Created and published :collector-exoplayer project."
 
 file="./bitmovin.properties"