+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - [Internal] `FeatureManager` allows dynamically adding and disabling features based on player support.
 - [Internal] `SegmentTrackingFeature` to always track the latest `n` downloaded elements
 - [Internal] `ErrorDetailsFeature` to track additional details in case an error occurs
+- [Internal] Enabled code style checks on pre-push and publish
 
 ## v1.22.0